Refactor challenge 'Sum Nums in Range'
 - Replace logic to sort array of two items with logic to use min and max of items in array

The Free Code Camp challenge offers helpful hints: Math.min(), Math.max(), and Array.reduce(). While use of these concepts is not specified as a requirement for completing the challenge, the challenge is designed to help students explore certain concepts. With this in mind, I refactored my solution to include the concepts offered in the hint section.